Automate updates to README
To improve the process of the README staying up-to-date, we can utilise
my new automagic package for generating the README, and pushing it to
GitLab whenever a scheduled build is triggered.

We can push via GitHub Actions' pre-build token, and want to make sure
that this only builds on scheduled executions.

We need to make sure that we install at least Go v1.17, as we require it
for `readme-generator`.

We also want to allow this to fail, as there may not be changes to
publish.